requirements.txt
fastapi
uvicorn
torch
faiss-cpu
pillow
numpy
git+https://github.com/openai/CLIP.git